Derivatives second commit
